+++ src/main/java/woojooin/planit/domain/goal/action/domain/Action.java
@@ -1,0 +1,32 @@
+package woojooin.planit.domain.goal.action.domain;
+
+import java.time.LocalDateTime;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class Action {
+
+	private Long actionId;
+
+	private Long goalId;
+
+	private String accountNumber;
+
+	private Long memberProductId;
+
+	private Integer allocatedRate;
+	private Integer accountAllocatedRate;
+
+	private String accountType;
+
+	private Integer amount;
+
+
+	private ActionType actionType;
+
+	private LocalDateTime createdAt;
+	private LocalDateTime updatedAt;
+}
