+++ src/main/java/woojooin/planit/domain/goal/isa/dto/req/IsaAccountProductEditReq.java
@@ -1,0 +1,23 @@
+package woojooin.planit.domain.goal.isa.dto.req;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class IsaAccountProductEditReq {
+
+	@NonNull
+	private Long goalId;
+
+	@NonNull
+	private Long memberProductId;
+
+	private String accountType = "ISA";
+
+	@NonNull
+	private boolean checked;
+}
