+++ src/main/java/woojooin/planit/domain/goal/isa/dto/res/IsaAccountTaxExemptionRes.java
@@ -1,0 +1,13 @@
+package woojooin.planit.domain.goal.isa.dto.res;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class IsaAccountTaxExemptionRes {
+
+	private long taxSavedAmount;
+}
