--- src/main/java/woojooin/planit/domain/object/action/domain/ActionType.java
@@ -1,25 +1,0 @@
-package woojooin.planit.domain.object.action.domain;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum ActionType {
-
-	ISA("ISA"),
-	DEPOSIT("예적금"),
-	;
-
-	private final String description;
-
-	public static ActionType fromCode(String actionType) {
-		for (ActionType type : ActionType.values()) {
-			if (type.getDescription().equals(actionType)) {
-				return type;
-			}
-		}
-
-		return null;
-	}
-}
