--- src/main/java/woojooin/planit/domain/object/deposit/dto/req/DepositAccountRegisterListReq.java
@@ -1,17 +1,0 @@
-package woojooin.planit.domain.object.deposit.dto.req;
-
-import java.util.List;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class DepositAccountRegisterListReq {
-
-	@NonNull
-	private List<DepositAccountRegisterReq> depositAccountRegisterReqs;
-}
