--- src/main/java/woojooin/planit/domain/object/isa/dto/req/IsaAccountProductEditReq.java
@@ -1,23 +1,0 @@
-package woojooin.planit.domain.object.isa.dto.req;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class IsaAccountProductEditReq {
-
-	@NonNull
-	private Long goalId;
-
-	@NonNull
-	private Long memberProductId;
-
-	private String accountType = "ISA";
-
-	@NonNull
-	private boolean checked;
-}
