+++ src/main/java/woojooin/planit/global/security/CustomUserDetailsService.java
@@ -1,5 +1,6 @@
 package woojooin.planit.global.security;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 import woojooin.planit.domain.member.domain.Member;
 import woojooin.planit.domain.member.repository.MemberRepository;
 
+@Slf4j
 @Service
 public class CustomUserDetailsService implements UserDetailsService {
 
@@ -20,6 +22,20 @@ public class CustomUserDetailsService implements UserDetailsService {
         Member member = memberRepository.findByEmail(username);
         if (member == null)
             throw new UsernameNotFoundException("User not found");
+        return new CustomUserDetails(member);
+    }
+
+    // 커스텀 메서드 추가 ✅
+    public UserDetails loadUserByMemberId(Long memberId) throws UsernameNotFoundException {
+        log.debug("Finding user by memberId: {}", memberId);
+
+        Member member = memberRepository.findById(memberId);
+        if (member == null) {
+            log.error("User not found with memberId: {}", memberId);
+            throw new UsernameNotFoundException("User not found with memberId: " + memberId);
+        }
+
+        log.debug("User found: {} (ID: {})", member.getMemberId(), member.getMemberId());
         return new CustomUserDetails(member);
     }
 }
