+++ src/main/java/woojooin/planit/global/security/jwt/JwtAuthenticationFilter.java
@@ -7,6 +7,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import io.jsonwebtoken.JwtException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -47,7 +48,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 		log.info("Processing JWT authentication for request: {}", request.getRequestURI());
 
 
-		if (request.getRequestURI().startsWith("/test") || request.getRequestURI().startsWith("/api/signup")) {
+		if ( request.getRequestURI().startsWith("/api")) {
 			log.info("Bypassing JWT filter for URI: {}", request.getRequestURI());
 			filterChain.doFilter(request, response);
 			return;
@@ -55,16 +56,19 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
 		String token = resolveToken(request);
 
-		if (token != null && jwtTokenProvider.validateToken(token)) {
+		if (token != null ) {
 			try {
+				jwtTokenProvider.validateToken(token);
+
 				String username = jwtTokenProvider.getUsername(token);
 				String role = jwtTokenProvider.getRole(token);
 				Long userId = jwtTokenProvider.getUserId(token);
 
 				Member member = memberService.findById(userId);
+
 				if (member == null) {
 					log.warn("Member not found for userId: {}", userId);
-					sendErrorResponse(response, "USER_NOT_FOUND", "사용자를 찾을 수 없습니다.");
+					sendErrorResponse(response, "USER_NOT_FOUND", "사용자를 찾을 수 없습니다!!!!!");
 					return;
 				}
 
@@ -83,15 +87,19 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 					return;
 				}
 
-				UserDetails userDetails = userDetailsService.loadUserByUsername(username);
+				UserDetails userDetails = userDetailsService.loadUserByMemberId(userId);
 
 				UsernamePasswordAuthenticationToken auth =
 						new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
 
 				SecurityContextHolder.getContext().setAuthentication(auth);
-			} catch (Exception e) {
+			} catch (JwtException e) {
 				log.error("JWT authentication error: {}", e.getMessage());
 				sendErrorResponse(response, "JWT_AUTH_ERROR", "JWT 인증 오류가 발생했습니다.");
+				return;
+			} catch (Exception e) {
+				log.error("Unexpected error during JWT authentication: {}", e.getMessage());
+				sendErrorResponse(response, "UNEXPECTED_ERROR", "예상치 못한 오류가 발생했습니다.");
 				return;
 			}
 		}
